+++ doc/CHANGES.rst
@@ -14,6 +14,9 @@ CHANGELOG
 Future
 ------
 
+0.2.7
+-----
+
 * [IMP] Add IPython support for interpreter
 * [REF] Update and Standardize the method to field the models (Field, Column, RelationShip)
   now all the type of the column go on the ftype and comme from the name of the class